Split gene network analysis from inference code
